+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Pending
+
+BREAKING CHANGES
+
+FEATURES
+
+IMPROVEMENTS
+* bank module uses go-wire codec instead of 'encoding/json'
+* auth module uses go-wire codec instead of 'encoding/json'
+
+FIXES
+
 ## 0.18.1
 
 BREAKING CHANGES